Move testing methods to new class
